# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HotCue {
  index: number;
  position: number;
  name: string;
  color: string;
}

export interface Track {
  id: string;
  fingerprint: string;
  name: string;
  artist: string;
  album: string;
  genre: string;
  duration: number;
  bpm?: number | null;
  bpmConfidence: number;
  key?: string | null;
  camelot?: string | null;
  energy: number;
  mood: string;
  danceability: number;
  loudness: number;
  albumArt?: string | null;
  isFavorite: boolean;
  colorLabel?: string | null;
  hotCues: HotCue[];
  loopIn?: number | null;
  loopOut?: number | null;
  createdAt: string;
  updatedAt: string;
}

/**
 * Same shape as Track, minus server-generated fields
 */
export type UpsertTrack = Track;

/**
 * Partial update, any subset of Track fields (excluding id/fingerprint)
 */
export interface UpdateTrack {
  name?: string;
  artist?: string;
  album?: string;
  genre?: string;
  duration?: number;
  bpm?: number | null;
  bpmConfidence?: number;
  key?: string | null;
  camelot?: string | null;
  energy?: number;
  mood?: string;
  danceability?: number;
  loudness?: number;
  albumArt?: string | null;
  isFavorite?: boolean;
  colorLabel?: string | null;
  hotCues?: HotCue[];
  loopIn?: number | null;
  loopOut?: number | null;
}