# lib/api-zod/src/generated/types/hotCue.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HotCue {
  index: number;
  position: number;
  name: string;
  color: string;
}