# lib/api-zod/src/generated/types/updateTrack.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HotCue } from './hotCue';

/**
 * Partial update, any subset of Track fields (excluding id/fingerprint)
 */
export interface UpdateTrack {
  name?: string;
  artist?: string;
  album?: string;
  genre?: string;
  duration?: number;
  bpm?: number | null;
  bpmConfidence?: number;
  key?: string | null;
  camelot?: string | null;
  energy?: number;
  mood?: string;
  danceability?: number;
  loudness?: number;
  albumArt?: string | null;
  isFavorite?: boolean;
  colorLabel?: string | null;
  hotCues?: HotCue[];
  loopIn?: number | null;
  loopOut?: number | null;
}