# lib/api-zod/src/generated/types/upsertTrack.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from './track';

/**
 * Same shape as Track, minus server-generated fields
 */
export type UpsertTrack = Track;